backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Web Scraping
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI Integration
anthropic==0.7.8
spacy==3.8.2
sentence-transformers==3.2.1
langgraph==0.2.53
python-docx==1.1.2
PyPDF2==3.0.1

# Scheduling
apscheduler==3.10.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.1

# Redis (for caching & task queue)
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# ASGI Server
uvloop==0.19.0
httptools==0.6.1
